requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.6.0
dbt-snowflake
apache-airflow-providers-snowflake==5.6.1
